refactor: 로또 주문 패키지 이동
