+++ src/main/java/lotto/domain/LottoInvest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.wrapper.LottoOrder;
 import lotto.view.Customer;
 
 public class LottoInvest {
